separate out testing code from the generation code
